+++ Examples/SDWebImage Demo/MasterViewController.m
@@ -124,8 +124,8 @@
         cell = [[MyCustomTableViewCell alloc] initWithStyle:UITableViewCellStyleDefault reuseIdentifier:CellIdentifier];
     }
 
-    [cell.customImageView setShowActivityIndicatorView:YES];
-    [cell.customImageView setIndicatorStyle:UIActivityIndicatorViewStyleGray];
+    [cell.customImageView sd_setShowActivityIndicatorView:YES];
+    [cell.customImageView sd_setIndicatorStyle:UIActivityIndicatorViewStyleGray];
     
     cell.customTextLabel.text = [NSString stringWithFormat:@"Image #%ld", (long)indexPath.row];
     [cell.customImageView sd_setImageWithURL:[NSURL URLWithString:_objects[indexPath.row]]